apps/agent/pyproject.toml
[tool.poetry.dependencies]
huggingface-hub = ">=0.33.2,<0.34.0"
langchain = ">=0.3.25,<0.4.0"
langchain-core = ">=0.3.68,<0.4.0"
langchain-exa = ">=0.3.0,<0.4.0"
langchain-openai = ">=0.3.16,<0.4.0"
langgraph = ">=0.4.3,<0.5.0"
langgraph-checkpoint-postgres = ">=2.0.22,<3.0.0"
langgraph-sdk = ">=0.1.69,<0.3.0"
nltk = ">=3.9.1,<4.0.0"
numpy = "<2"
pydantic-settings = ">=2.10.1,<3.0.0"
python = ">=3.11,<4.0"
python-dotenv = ">=1.1.0,<2.0.0"
redis = "^6.2.0"
scikit-learn = ">=1.7.0,<2.0.0"
scipy = ">=1.16.0,<2.0.0"
sentence-transformers = ">=5.0.0,<6.0.0"
torch = ">=2.7.1,<3.0.0"
transformers = ">=4.53.1,<5.0.0"
uvicorn = "^0.35.0"
starlette = ">=0.22.0,<1.0.0"
langchain-tavily = "^0.2.9"

[tool.poetry]
authors = ["Your Name <your.email@example.com>"]
description = "Development environment for fact-checking agents"
name = "fact-checker-agent"
packages = [
  {include = "claim_extractor"},
  {include = "claim_verifier"},
  {include = "fact_checker"},
  {include = "utils"},
]
version = "0.1.0"

[tool.poetry.scripts]
create-run = "scripts.create_run:main"
dev = "scripts.dev:main"
api-key = "scripts.api_key:main"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
